refactor: isolate OpenCode plugin adapter
